Fixed Employee thing. Postman connecting to db

+++ src/main/java/dev/evanishyn/handlers/employeeHandlers/DeleteEmployeeHandler.java
@@ -1,6 +1,7 @@
 package dev.evanishyn.handlers.employeeHandlers;
 
 import dev.evanishyn.app.App;
+import dev.evanishyn.entities.Employee;
 import io.javalin.http.Context;
 import io.javalin.http.Handler;
 import org.jetbrains.annotations.NotNull;
@@ -9,15 +10,15 @@ public class DeleteEmployeeHandler implements Handler {
 
     @Override
     public void handle(@NotNull Context ctx) throws Exception{
-        int id = Integer.parseInt(ctx.pathParam("id"));
-        boolean result = App.employeeService.deleteEmployeeAccount(id);
-        if(result){
-            ctx.status(204);
-        }else{
+        int id = Integer.parseInt(ctx.pathParam("emp_id"));
+        Employee employee =App.employeeService.getEmployeeById(id);
+        if(employee == null){
             ctx.status(404);
             ctx.result("Employee not found");
+        } else {
+            App.employeeService.deleteEmployeeAccount(id);
+            ctx.status(204);
+            ctx.result("Employee has been deleted");
         }
-
     }
-
 }
